test: Fix cast for Windows 64-bit.
* tests/bench-slope.c (slope_benchmark): Use uintptr_t.

--

Fixes-commit: 1517a31ea476b76f67289d07e2c2821aa4ef83ef

+++ tests/bench-slope.c
@@ -467,7 +467,7 @@ slope_benchmark (struct bench_obj *obj)
     goto err_free;
   /* Get aligned buffer */
   buffer = real_buffer;
-  buffer += 128 - ((unsigned long)real_buffer & (128 - 1));
+  buffer += 128 - ((uintptr_t)real_buffer & (128 - 1));
   if (unaligned_mode)
     buffer += unaligned_mode; /* Make buffer unaligned */
